% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionFunctions.R
\name{EWASex.getGoldCpGNames}
\alias{EWASex.getGoldCpGNames}
\title{Get the 49 goldset CpGs.}
\usage{
EWASex.getGoldCpGNames()
}
\value{
A character vector with the names of the 49 CpGs that best predict sex.
}
\description{
Allows for easy subsetting of a data.frame.
}
\examples{
CpGs <- EWASex.getGoldCpGNames()

}
