% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionFunctions.R
\name{EWASex.predict}
\alias{EWASex.predict}
\title{Computes the errors, normalized errors and performs the sex predictions.}
\usage{
EWASex.predict(df, means, margin = 2, use_normalized = TRUE)
}
\arguments{
\item{df}{The data.frame consiting of the CpG information. This can be any CpGs, but it is recommended to use the gold set of 49 CpGs.}

\item{means}{Means and SDs already computed using the \code{getMeansAndSD} function.}

\item{margin}{Whether to compute the means and SDs column or row-wise. Default is 2, corresponding to one sample per row and a CpG for each column.}

\item{use_normalized}{If you expect or know that your dataset would only consists of a single sex (e.g. only females) then change this to FALSE, as normaliztion wouldn't work as expected on a single sex.}
}
\value{
The means and SDs of \code{df} for each gender in \code{genders}
}
\description{
Computes the errors, normalized errors and performs the sex predictions.
}
\examples{
# Example:
# if you already have means and SD from the current of another dataset
df <- wholeDataFrame[,EWASex.getGoldCpGNames()] # subset to only use the 49 CpGs
data("MeansAndSD49")

predictions <- EWASex.predict(df, means=MeansAndSD49) # Alternative, \code{means} can be computed with \code{EWASex.train}
}
