% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionFunctions.R
\name{EWASex.train}
\alias{EWASex.train}
\title{Computes the vectors for the means and SDs that are used for the prediction part.}
\usage{
EWASex.train(genders, df, margin = 2)
}
\arguments{
\item{genders}{Gender information indicating the genders. This can be a vector consiting of 1's and 2's,
or characters, like, "male" and "female".}

\item{df}{The data.frame consiting of the CpG information. This can be any CpGs, but it is recommended to use the gold set of 49 CpGs.}

\item{margin}{Whether to compute the means and SDs column or row-wise. Default is 2, corresponding to one sample per row and a CpG for each column.}
}
\value{
The means and SDs of \code{df} for each gender in \code{genders}
}
\description{
The vectors can be stored to predict at a later time or used on another dataset that doesn't contain gender information.
If you are only intrested in predicting gender, you can use \code{getPredictions} with \code{means} = "none".
}
\examples{
genders <- pheno$genders # get genders from your phenotype file
df <- wholeDataFrame[,EWASex.getGoldCpGNames()] # subset to only use the 49 CpGs

MeansAndSD <- EWASex.train(genders, df, margin=2)
}
