% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionFunctions.R
\docType{data}
\name{MeansAndSD49}
\alias{MeansAndSD49}
\title{Precomputed means and SD from 49 CpGs.}
\format{
An object of class \code{"data.frame"}; The keys represents genders, here 1 represent a "male" and 2 represent a "female".
}
\source{
Combined means and SDs computed from Middle Aged Danish Twins (MADT),
Longitudinal Study of Aging Danish Twins 2 (LSADT2), and, Birth-weight Discordant Twins (BWT).
(See https://doi.org/10.1017/thg.2012.77 and https://doi.org/10.1017/thg.2012.77 for dataset descriptions)
}
\usage{
data(MeansAndSD49)
}
\description{
Means and SD from 49 CpGs, computed from MADT, BWT, and LSADT2 cohorts. See @source.
}
\examples{
data(MeansAndSD49)

}
\references{
Not published yet.
}
\keyword{datasets}
